requirements.txt
click==8.1.6
krfzf-py==0.0.4
